+++ app/app_upgrades_test.go
@@ -7,10 +7,10 @@ import (
 	storeTypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
-	"github.com/cosmos/cosmos-sdk/x/auth/keeper"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 	"github.com/stretchr/testify/require"
 
+	"github.com/archway-network/archway/app/keepers"
 	"github.com/archway-network/archway/app/upgrades"
 	e2eTesting "github.com/archway-network/archway/e2e/testing"
 	rewardsTypes "github.com/archway-network/archway/x/rewards/types"
@@ -45,7 +45,7 @@ func TestUpgrades(t *testing.T) {
 	proposalExecuted := false
 	fauxUpgrade := upgrades.Upgrade{
 		UpgradeName: "test-upgrade",
-		CreateUpgradeHandler: func(manager *module.Manager, configurator module.Configurator, _ keeper.AccountKeeper) upgradetypes.UpgradeHandler {
+		CreateUpgradeHandler: func(manager *module.Manager, configurator module.Configurator, _ keepers.ArchwayKeepers) upgradetypes.UpgradeHandler {
 			return func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 				proposalExecuted = true
 				return manager.RunMigrations(ctx, configurator, fromVM)