+++ x/rewards/keeper/state_flatfee_test.go
@@ -10,7 +10,7 @@ import (
 // TestFlatFeeImportExport check flat fees import/export.
 // Test updates the initial state with new records and checks that they were merged.
 func (s *KeeperTestSuite) TestFlatFeeImportExport() {
-	ctx, keeper := s.chain.GetContext(), s.chain.GetApp().RewardsKeeper
+	ctx, keeper := s.chain.GetContext(), s.chain.GetApp().Keepers.RewardsKeeper
 	contractAddrs := e2eTesting.GenContractAddresses(2)
 
 	newFlatFees := []rewardsTypes.FlatFee{